+++ custom_components/saj_h2_modbus/const.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 from typing import Any, Optional, Literal, TypeAlias
 from dataclasses import dataclass
@@ -83,7 +85,6 @@ power_sensors_group = SensorGroup(
     device_class=SensorDeviceClass.POWER,
     state_class=SensorStateClass.MEASUREMENT,
     icon="mdi:solar-power",
-    force_update=True,  # enable force_update for the entire group
 )
 
 apparent_power_sensors_group = SensorGroup(
@@ -91,7 +92,6 @@ apparent_power_sensors_group = SensorGroup(
     device_class=SensorDeviceClass.APPARENT_POWER,
     state_class=SensorStateClass.MEASUREMENT,
     icon="mdi:flash-outline",
-    force_update=True,
 )
 
 voltage_sensors_group = SensorGroup(